Add CoverImagesList + responsiveness
